[googlemaps] Extract data.d.ts
